Fix the ugly warranty script
I dunno what the hell I was thinking, but I tried to parse JSON
with a bunch of splits and regex...which was ugly as hell. Now the
script ACTUALLY USES the JSON library to parse a JSON string (weird,
huh?). Feel free to take this and run with it.