Fix for Joe Chilicote
So, it was failing in certain cases when rubygems wasn't required. There
are also pieces of hardware out there that don't have a COV_END_DATE but
were still under warranty coverage. For those cases, I use the
HW_END_DATE and then convert it to a nicely displayed date.