Fix warranty script...again
Fucking hell, Apple.  You felt the need to NOT return serialized data
and instead simply present a pretty web page.  This is all well and good
for people visiting your site, but it would be nice to be able to
collect this data in some presentable manner.  Thanks to frogor from
\#\#osx-server (https://github.com/pudquick) for running down the changes
during the day.  You da man!